+++ Changelog.md
@@ -1,11 +1,17 @@
+# v0.2.0 (2013-31-03)
+
+#### Breaking Changes:
+
+- API change: create instances of EditableJS [65](https://github.com/upfrontIO/Editable.JS/pull/65)
+
 
 # v0.1.2 (2014-03-07)
 
 - [Add selection methods](https://github.com/upfrontIO/Editable.JS/pull/64)
-  - New Selection methods:  
-    #collapseAtBeginning()  
+  - New Selection methods:
+    #collapseAtBeginning()
     #collapseAtEnd()
-  - New Cursor and Selection method:  
+  - New Cursor and Selection method:
     #setVisibleSelection() (alias for #setSelection())
 
 # v0.1.1 (2014-01-24)